fix some duplicate calls, and add some logs

+++ webgl/initBuffers.js
@@ -1,4 +1,4 @@
-import { CUBE_POSITIONS, positions } from './constants.js';
+import { positions } from './constants.js';
 import { setColors } from './setColors.js';
 
 const createPositionBuffer = (gl) => {
@@ -57,16 +57,17 @@ const createCubeColorBuffers = (gl) => {
     //     16, 17, 18, 16, 18, 19, // right
     //     20, 21, 22, 20, 22, 23,  // left
     // ];
-      const indices = [
-    0,  1,  2,      0,  2,  3,    // front
-    4,  5,  6,      4,  6,  7,    // back
-    8,  9,  10,     8,  10, 11,   // top
-    12, 13, 14,     12, 14, 15,   // bottom
-    16, 17, 18,     16, 18, 19,   // right
-    20, 21, 22,     20, 22, 23,   // left
-  ];
+    const indices = [
+        0,  1,  2,      0,  2,  3,    // front
+        4,  5,  6,      4,  6,  7,    // back
+        8,  9,  10,     8,  10, 11,   // top
+        12, 13, 14,     12, 14, 15,   // bottom
+        16, 17, 18,     16, 18, 19,   // right
+        20, 21, 22,     20, 22, 23,   // left
+    ];
 
-    gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, new Uint8Array(indices), gl.STATIC_DRAW);
+    // TODO: why does this matter usin ga Uint8Array vs a Uint16Array (?)
+    gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, new Uint16Array(indices), gl.STATIC_DRAW);
     return indexBuffer;
 }
